integrating flask_saas and adding requirements
