+++ requirements.txt
@@ -2,3 +2,5 @@ Flask==2.0.1
 python-dotenv==0.19.0
 stripe==2.60.0
 livereload==2.5.1
+git+https://github.com/karmacomputing/flask-saas.git@v0.0.4
+
